Update ABRAhaM_CONFIGUI.py
Updated Configui.py script.  Accounts section both add and edit are in a working state.  Preferences section is only about 1/3rd done.  ConfigFile operations about 1/3rd as well but fileops should be small and easy in this script.  Will be working on the preferences section next and since a lot of that coding will be nearly the same as the accounts work it should go fast.